Added Daily Report and Monthly Category report. Changed from Long to double

+++ smf-domain/src/main/java/com/smf/main/domain/FundDao.java
@@ -21,5 +21,5 @@ public interface FundDao extends JpaRepository<Fund, Long> {
     Long findFundByUser(Long fundId, String userName);
 
     @Query(value = "select sum(f.amount) from fund f where f.fk_user = ?1", nativeQuery = true)
-    Long findTotalAmountForUser(Long userId);
+    Double findTotalAmountForUser(Long userId);
 }
